Mostly working version of the core generation engine

+++ include/robin/color/pixel_accumulation.h
@@ -1,8 +1,10 @@
 #pragma once
 
+#include <cstdint>
+
 namespace ff {
   struct PixelAccumulation {
-    float frequency_{};
+    int frequency_{};
     float red_{};
     float green_{};
     float blue_{};
